requirements.txt
seaborn==0.11.1
numpy==1.20.3
pandas==1.2.4
streamlit==0.62.1
Pillow==8.2.0
